app/requirements.txt
diffusers==0.8.0
transformers
# scipy 
ftfy
accelerate
huggingface_hub
# torch
# absl-py==1.3.0
anyio==3.6.2
# astunparse==1.6.3
# cachetools==5.2.0
# certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_0ek9yztvu3/croot/certifi_1665076692562/work/certifi
# certifi
# charset-normalizer==2.1.1
# click==8.1.3
fastapi==0.88.0
# flatbuffers==22.11.23
# gast==0.4.0
# google-auth==2.14.1
# google-auth-oauthlib==0.4.6
# google-pasta==0.2.0
# grpcio==1.50.0
# h11==0.14.0
# h5py==3.7.0
# idna==3.4
# keras==2.11.0
# libclang==14.0.6
# Markdown==3.4.1
# MarkupSafe==2.1.1
nest-asyncio==1.5.6
# numpy==1.23
# oauthlib==3.2.2
# opt-einsum==3.3.0
# packaging==21.3
Pillow==9.3.0
# protobuf==3.19.6
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pydantic==1.10.2
# pyparsing==3.0.9
python-multipart==0.0.5
# requests==2.28.1
# requests-oauthlib==1.3.1
# rsa==4.9
# six==1.16.0
# sniffio==1.3.0
# starlette==0.22.0
# tensorboard==2.11.0
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorflow==2.11.0
# tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
# tensorflow-io-gcs-filesystem==0.28.0
# termcolor==2.1.1
# typing_extensions==4.4.0
# urllib3==1.26.13
uvicorn==0.20.0
# Werkzeug==2.2.2
# wrapt==1.14.1